move jimm interface to a separate file
